requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.6
google-genai==1.64.0
fastapi==0.129.2
uvicorn[standard]==0.41.0
python-multipart==0.0.22
python-dotenv==1.1.1
